+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.1.2
+
+- Refactor settings to ``conf.py``.
+
 ## 0.1.1
 
 - Refactor adding wspay payment.